+++ backend/src/Logitar.Portal.Application/Users/Commands/RemoveUserIdentifierCommand.cs
@@ -1,7 +1,38 @@
-﻿using Logitar.Portal.Application.Activities;
+﻿using Logitar.EventSourcing;
+using Logitar.Identity.Domain.Users;
+using Logitar.Portal.Application.Activities;
 using Logitar.Portal.Contracts.Users;
 using MediatR;
 
 namespace Logitar.Portal.Application.Users.Commands;
 
 internal record RemoveUserIdentifierCommand(Guid Id, string Key) : Activity, IRequest<UserModel?>;
+
+internal class RemoveUserIdentifierCommandHandler : IRequestHandler<RemoveUserIdentifierCommand, UserModel?>
+{
+  private readonly IUserManager _userManager;
+  private readonly IUserQuerier _userQuerier;
+  private readonly IUserRepository _userRepository;
+
+  public RemoveUserIdentifierCommandHandler(IUserManager userManager, IUserQuerier userQuerier, IUserRepository userRepository)
+  {
+    _userManager = userManager;
+    _userQuerier = userQuerier;
+    _userRepository = userRepository;
+  }
+
+  public async Task<UserModel?> Handle(RemoveUserIdentifierCommand command, CancellationToken cancellationToken)
+  {
+    UserAggregate? user = await _userRepository.LoadAsync(command.Id, cancellationToken);
+    if (user == null || user.TenantId != command.TenantId)
+    {
+      return null;
+    }
+
+    ActorId actorId = command.ActorId;
+    user.RemoveCustomIdentifier(command.Key, actorId);
+    await _userManager.SaveAsync(user, command.UserSettings, actorId, cancellationToken);
+
+    return await _userQuerier.ReadAsync(command.Realm, user, cancellationToken);
+  }
+}
