+++ backend/src/Logitar.Portal.Application/Users/Commands/SignOutUserCommand.cs
@@ -1,7 +1,44 @@
-﻿using Logitar.Portal.Application.Activities;
+﻿using Logitar.EventSourcing;
+using Logitar.Identity.Domain.Sessions;
+using Logitar.Identity.Domain.Users;
+using Logitar.Portal.Application.Activities;
 using Logitar.Portal.Contracts.Users;
 using MediatR;
 
 namespace Logitar.Portal.Application.Users.Commands;
 
 public record SignOutUserCommand(Guid Id) : Activity, IRequest<UserModel?>;
+
+internal class SignOutUserCommandHandler : IRequestHandler<SignOutUserCommand, UserModel?>
+{
+  private readonly ISessionRepository _sessionRepository;
+  private readonly IUserQuerier _userQuerier;
+  private readonly IUserRepository _userRepository;
+
+  public SignOutUserCommandHandler(ISessionRepository sessionRepository, IUserQuerier userQuerier, IUserRepository userRepository)
+  {
+    _sessionRepository = sessionRepository;
+    _userQuerier = userQuerier;
+    _userRepository = userRepository;
+  }
+
+  public async Task<UserModel?> Handle(SignOutUserCommand command, CancellationToken cancellationToken)
+  {
+    UserAggregate? user = await _userRepository.LoadAsync(command.Id, cancellationToken);
+    if (user == null || user.TenantId != command.TenantId)
+    {
+      return null;
+    }
+
+    ActorId actorId = command.ActorId;
+    IEnumerable<SessionAggregate> sessions = await _sessionRepository.LoadActiveAsync(user, cancellationToken);
+    foreach (SessionAggregate session in sessions)
+    {
+      session.SignOut(actorId);
+    }
+
+    await _sessionRepository.SaveAsync(sessions, cancellationToken);
+
+    return await _userQuerier.ReadAsync(command.Realm, user, cancellationToken);
+  }
+}
