--- backend/src/Logitar.Portal.Application/Users/Commands/SignOutUserCommandHandler.cs
@@ -1,41 +1,0 @@
-﻿using Logitar.EventSourcing;
-using Logitar.Identity.Domain.Sessions;
-using Logitar.Identity.Domain.Users;
-using Logitar.Portal.Contracts.Users;
-using MediatR;
-
-namespace Logitar.Portal.Application.Users.Commands;
-
-internal class SignOutUserCommandHandler : IRequestHandler<SignOutUserCommand, UserModel?>
-{
-  private readonly ISessionRepository _sessionRepository;
-  private readonly IUserQuerier _userQuerier;
-  private readonly IUserRepository _userRepository;
-
-  public SignOutUserCommandHandler(ISessionRepository sessionRepository, IUserQuerier userQuerier, IUserRepository userRepository)
-  {
-    _sessionRepository = sessionRepository;
-    _userQuerier = userQuerier;
-    _userRepository = userRepository;
-  }
-
-  public async Task<UserModel?> Handle(SignOutUserCommand command, CancellationToken cancellationToken)
-  {
-    UserAggregate? user = await _userRepository.LoadAsync(command.Id, cancellationToken);
-    if (user == null || user.TenantId != command.TenantId)
-    {
-      return null;
-    }
-
-    ActorId actorId = command.ActorId;
-    IEnumerable<SessionAggregate> sessions = await _sessionRepository.LoadActiveAsync(user, cancellationToken);
-    foreach (SessionAggregate session in sessions)
-    {
-      session.SignOut(actorId);
-    }
-
-    await _sessionRepository.SaveAsync(sessions, cancellationToken);
-
-    return await _userQuerier.ReadAsync(command.Realm, user, cancellationToken);
-  }
-}
